Temperately hide dyn blogs
There're too many bugs with Layout + metadata API in the latest version
of Next.JS

Currently CSS imported in non-root layout will not be included in
production build. The bug is known by developers.

I'm just too tired to fix it for them or create a workaround. Though the
workaround should be easy: import the CSS in root layout.

So temperately, I'm hiding the Dyn page from you.